[rv32i] Enhancement: Refactoring Registers Header
